Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3: Code-Review+2
> 
> Thank you! This also now includes the case '...--><!--...'.


Just a question to the tests in xt/benchmark.pl: The new test (delHTMLcom-long) does 10.000 iterations, while the old test (delHTMLcom) does 100.000 iterations. Shouldn't we use the same number of iterations for comparibility?
The concatenation with $_ in the test 'delHTMLcom' is useless (or not?). The specification *STDIN is o.k., because it's not read (otherwise it would stuck/fail). The alternative is to also use a filehandle, which may be overdone ...

Patch-set: 3
Reviewer: Gerrit User 25 <25@626241b7-b7d0-4197-b843-1ee23887649c>