Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3: Code-Review+2
> > 
> > Thank you! This also now includes the case '...--><!--...'.
> 
> 
> Just a question to the tests in xt/benchmark.pl: The new test (delHTMLcom-long) does 10.000 iterations, while the old test (delHTMLcom) does 100.000 iterations. Shouldn't we use the same number of iterations for comparibility?

Different benchmarks can't be compared - only different runs of the same benchmark. The benchmarks automatically run an arbitrary number of times until the algorithm decides the values are stable - so principally there is no need to add a "for" loop for multiple inner runs. But I prefer single benchmarks to take at least microseconds, so I understand how long it really took and if a slowdown is actually meaningful or neglectable. That's just the only reason for the loop. ;) I am fine with removing the loops altogether, but then we should render the values in a more human readable way.

> The concatenation with $_ in the test 'delHTMLcom' is useless (or not?).

For benchmarking it is always suggested to change parameters for the calls to benchmark, because there is no guarantee, that the compiler doesn't "cache" the result (because the compiler may be smart enough to see that a function without inner or external state and called with the same parameters should always return the same value). Before I check if a function has any state, I always prefer to modify the parameters to call.
It's also suggested to return the result (`$result`), so the compiler doesn't clean up the AST from function calls that won't have any actual effects on the program (nor side effects).

> The specification *STDIN is o.k., because it's not read (otherwise it would stuck/fail). The alternative is to also use a filehandle, which may be overdone ...

Yes. I don't know if it is relevant in the benchmark.

Patch-set: 3
Reviewer: Gerrit User 19 <19@626241b7-b7d0-4197-b843-1ee23887649c>