Update patch set 4
Patch Set 4:

I found a bug within the handling of the intern tokenization (conservative, aggressive) in script/tei2korapxml, inside the function retr_info(), where $_GEN_TOK_INT is checked.

Because of this bug, text nodes starting with 2 blanks are omitted during the intern tokenization.

I added 2 new tests to t/script.t, which will fail when tested in this patch.

The solution/bugfix is to simply remove an (anyway unnecessary) if-clause inside retr_info() - which is described in patchset 1.

When commenting out the if-clause, the testing for this patch will be successfull.

I will do that in the next patch, also cleaning up comments a little bit.

Patch-set: 4