Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > Patch Set 9:
> > 
> > > Patch Set 9:
> > > 
> > > > Patch Set 7:
> > > > 
> > > > Do you still try to write a test for the produced data or should we merge and see, if any problems occur?
> > > 
> > > I corrected t/script.t and adjusted script/tei2korapxml, so that the subtest 'Test utf-8 handling' runs without errors.
> > > I'm still not so sure about the different possibilities of output for non-utf8 codepoints, but we can change the subtest again after merging, if necessary ;)
> > 
> > I did the change in Patchset 5
> 
> I expected you would check the output of the generated file regarding validity.

Ah, ok, thank you for the note the validation of the temporary zip-file (with "script_out5" in its name).
It contains the sequence "þ<83>¿¿¿¿¿" in all path-names and also in the 'docid'-fields, which is the expected output, when using lax utf8-encoding.
All 'sigle'-fields (in the header files) contain the sequence "Ã¾Â<83>Â¿Â¿Â¿Â¿Â¿", which is the result of encoding the already utf8-encode string again.
This is not expected and should be an error in the processing of the header files.

Shall we first merge, or correct it in this commit?

Patch-set: 9
Attention: {"person_ident":"Gerrit User 25 \u003c25@626241b7-b7d0-4197-b843-1ee23887649c\u003e","operation":"REMOVE","reason":"removed on reply"}
Attention: {"person_ident":"Gerrit User 19 \u003c19@626241b7-b7d0-4197-b843-1ee23887649c\u003e","operation":"ADD","reason":"owner or uploader replied"}