Update patch set 6
Patch Set 6:

(10 comments)

This removes quite a lot of code, as it turns out: Tokens and Structures are nearly identical, and when reusing the annotation for both collections, a lot of changes, bookkeeping etc. can be removed completely. However - because of the limited tests on structures, there may be edge cases that need to be tested.

Patch-set: 6