Update patch set 1
Patch Set 1:

Maybe thats a controversial commit.

Especially for code reviews and optimization/refactoring the long and verbose documentation sections in the code are imho quite distracting.

Hints what happens when a line is removed (helpful for refactoring) should now be covered by tests instead, so are not helpful anymore.

I guess also comments like "This starts here ... until here" are no longer necessary when the code is shorter (as it is now already after the modularization).

Patch-set: 1