Update patch set 1
Patch Set 1:

I think this is a cleaner way of skipping the sentences without introducing and decorating a dummy annotation.

Maybe we should add a "--skip-annotation" parameter where "s" will be added once sentence-splits are required. Then we wouldn't have to hard-code "s" (giving the flexibility to work and skip other annotations as well).

Patch-set: 1