Update patch set 1
Patch Set 1:

This is a test suite regarding the whitespace handling following the documentation by Peter. While we carefully tried to not change the original behaviour of the script, the documented behaviour is not what the script currently does.

I think it's not too complicated to fix that and make it work following the documentation, but a) we should add better regression tests, it seems and b) I don't really understand the intended behaviour, so maybe someone can explain th documentation to me a bit more.

Patch-set: 1